+++ cmake/templates/opencv_abi.xml.in
@@ -23,6 +23,7 @@
 <skip_headers>
     opencv2/core/hal/intrin*
     opencv2/core/cuda*
+    opencv2/core/opencl*
     opencv2/core/private*
     opencv/cxeigen.hpp
     opencv2/core/eigen.hpp
