+++ modules/ts/src/ocl_test.cpp
@@ -43,16 +43,9 @@
 
 #include "opencv2/ts/ocl_test.hpp"
 
-namespace cvtest {
-namespace ocl {
-
-using namespace cv;
-
-int test_loop_times = 1; // TODO Read from command line / environment
-
 #ifdef HAVE_OPENCL
 
-#define DUMP_PROPERTY_XML(propertyName, propertyValue) \
+#define DUMP_CONFIG_PROPERTY(propertyName, propertyValue) \
     do { \
         std::stringstream ssName, ssValue;\
         ssName << propertyName;\
@@ -65,180 +58,23 @@ int test_loop_times = 1; // TODO Read from command line / environment
         std::cout << msg << std::endl; \
     } while (false)
 
-static std::string bytesToStringRepr(size_t value)
-{
-    size_t b = value % 1024;
-    value /= 1024;
+#include <opencv2/core/opencl/opencl_info.hpp>
 
-    size_t kb = value % 1024;
-    value /= 1024;
+#endif // HAVE_OPENCL
 
-    size_t mb = value % 1024;
-    value /= 1024;
+namespace cvtest {
+namespace ocl {
 
-    size_t gb = value;
+using namespace cv;
 
-    std::ostringstream stream;
+int test_loop_times = 1; // TODO Read from command line / environment
 
-    if (gb > 0)
-        stream << gb << " GB ";
-    if (mb > 0)
-        stream << mb << " MB ";
-    if (kb > 0)
-        stream << kb << " kB ";
-    if (b > 0)
-        stream << b << " B";
-
-    return stream.str();
-}
-
+#ifdef HAVE_OPENCL
 void dumpOpenCLDevice()
 {
-    using namespace cv::ocl;
-
-    try
-    {
-        if (!useOpenCL())
-        {
-            DUMP_MESSAGE_STDOUT("OpenCL is disabled");
-            DUMP_PROPERTY_XML("cv_ocl", "disabled");
-            return;
-        }
-
-        std::vector<PlatformInfo> platforms;
-        cv::ocl::getPlatfomsInfo(platforms);
-        if (platforms.size() > 0)
-        {
-            DUMP_MESSAGE_STDOUT("OpenCL Platforms: ");
-            for (size_t i = 0; i < platforms.size(); i++)
-            {
-                const PlatformInfo* platform = &platforms[i];
-                DUMP_MESSAGE_STDOUT("    " << platform->name().c_str());
-                Device current_device;
-                for (int j = 0; j < platform->deviceNumber(); j++)
-                {
-                    platform->getDevice(current_device, j);
-                    const char* deviceTypeStr = current_device.type() == Device::TYPE_CPU
-                        ? ("CPU") : (current_device.type() == Device::TYPE_GPU ? current_device.hostUnifiedMemory() ? "iGPU" : "dGPU" : "unknown");
-                    DUMP_MESSAGE_STDOUT( "        " << deviceTypeStr << ": " << current_device.name().c_str() << " (" << current_device.version().c_str() << ")");
-                    DUMP_PROPERTY_XML( cv::format("cv_ocl_platform_%d_device_%d", (int)i, (int)j ),
-                        cv::format("(Platform=%s)(Type=%s)(Name=%s)(Version=%s)",
-                        platform->name().c_str(), deviceTypeStr, current_device.name().c_str(), current_device.version().c_str()) );
-                }
-            }
-        }
-        else
-        {
-            DUMP_MESSAGE_STDOUT("OpenCL is not available");
-            DUMP_PROPERTY_XML("cv_ocl", "not available");
-            return;
-        }
-
-        const Device& device = Device::getDefault();
-        if (!device.available())
-            CV_ErrorNoReturn(CV_OpenCLInitError, "OpenCL device is not available");
-
-        DUMP_MESSAGE_STDOUT("Current OpenCL device: ");
-
-#if 0
-        DUMP_MESSAGE_STDOUT("    Platform = "<< device.getPlatform().name());
-        DUMP_PROPERTY_XML("cv_ocl_current_platformName", device.getPlatform().name());
-#endif
-
-        const char* deviceTypeStr = device.type() == Device::TYPE_CPU
-            ? ("CPU") : (device.type() == Device::TYPE_GPU ? device.hostUnifiedMemory() ? "iGPU" : "dGPU" : "unknown");
-        DUMP_MESSAGE_STDOUT("    Type = "<< deviceTypeStr);
-        DUMP_PROPERTY_XML("cv_ocl_current_deviceType", deviceTypeStr);
-
-        DUMP_MESSAGE_STDOUT("    Name = "<< device.name());
-        DUMP_PROPERTY_XML("cv_ocl_current_deviceName", device.name());
-
-        DUMP_MESSAGE_STDOUT("    Version = " << device.version());
-        DUMP_PROPERTY_XML("cv_ocl_current_deviceVersion", device.version());
-
-        DUMP_MESSAGE_STDOUT("    Driver version = " << device.driverVersion());
-        DUMP_PROPERTY_XML("cv_ocl_current_driverVersion", device.driverVersion());
-
-        DUMP_MESSAGE_STDOUT("    Address bits = " << device.addressBits());
-        DUMP_PROPERTY_XML("cv_ocl_current_addressBits", device.addressBits());
-
-        DUMP_MESSAGE_STDOUT("    Compute units = "<< device.maxComputeUnits());
-        DUMP_PROPERTY_XML("cv_ocl_current_maxComputeUnits", device.maxComputeUnits());
-
-        DUMP_MESSAGE_STDOUT("    Max work group size = "<< device.maxWorkGroupSize());
-        DUMP_PROPERTY_XML("cv_ocl_current_maxWorkGroupSize", device.maxWorkGroupSize());
-
-        std::string localMemorySizeStr = bytesToStringRepr(device.localMemSize());
-        DUMP_MESSAGE_STDOUT("    Local memory size = " << localMemorySizeStr);
-        DUMP_PROPERTY_XML("cv_ocl_current_localMemSize", device.localMemSize());
-
-        std::string maxMemAllocSizeStr = bytesToStringRepr(device.maxMemAllocSize());
-        DUMP_MESSAGE_STDOUT("    Max memory allocation size = "<< maxMemAllocSizeStr);
-        DUMP_PROPERTY_XML("cv_ocl_current_maxMemAllocSize", device.maxMemAllocSize());
-
-        const char* doubleSupportStr = device.doubleFPConfig() > 0 ? "Yes" : "No";
-        DUMP_MESSAGE_STDOUT("    Double support = "<< doubleSupportStr);
-        DUMP_PROPERTY_XML("cv_ocl_current_haveDoubleSupport", device.doubleFPConfig() > 0);
-
-        const char* isUnifiedMemoryStr = device.hostUnifiedMemory() ? "Yes" : "No";
-        DUMP_MESSAGE_STDOUT("    Host unified memory = "<< isUnifiedMemoryStr);
-        DUMP_PROPERTY_XML("cv_ocl_current_hostUnifiedMemory", device.hostUnifiedMemory());
-
-        DUMP_MESSAGE_STDOUT("    Device extensions:");
-        String extensionsStr = device.extensions();
-        size_t pos = 0;
-        while (pos < extensionsStr.size())
-        {
-            size_t pos2 = extensionsStr.find(' ', pos);
-            if (pos2 == String::npos)
-                pos2 = extensionsStr.size();
-            if (pos2 > pos)
-            {
-                String extensionName = extensionsStr.substr(pos, pos2 - pos);
-                DUMP_MESSAGE_STDOUT("        " << extensionName);
-            }
-            pos = pos2 + 1;
-        }
-        DUMP_PROPERTY_XML("cv_ocl_current_extensions", extensionsStr.c_str());
-
-        const char* haveAmdBlasStr = haveAmdBlas() ? "Yes" : "No";
-        DUMP_MESSAGE_STDOUT("    Has AMD Blas = "<< haveAmdBlasStr);
-        DUMP_PROPERTY_XML("cv_ocl_current_AmdBlas", haveAmdBlas());
-
-        const char* haveAmdFftStr = haveAmdFft() ? "Yes" : "No";
-        DUMP_MESSAGE_STDOUT("    Has AMD Fft = "<< haveAmdFftStr);
-        DUMP_PROPERTY_XML("cv_ocl_current_AmdFft", haveAmdFft());
-
-
-        DUMP_MESSAGE_STDOUT("    Preferred vector width char = "<< device.preferredVectorWidthChar());
-        DUMP_PROPERTY_XML("cv_ocl_current_preferredVectorWidthChar", device.preferredVectorWidthChar());
-
-        DUMP_MESSAGE_STDOUT("    Preferred vector width short = "<< device.preferredVectorWidthShort());
-        DUMP_PROPERTY_XML("cv_ocl_current_preferredVectorWidthShort", device.preferredVectorWidthShort());
-
-        DUMP_MESSAGE_STDOUT("    Preferred vector width int = "<< device.preferredVectorWidthInt());
-        DUMP_PROPERTY_XML("cv_ocl_current_preferredVectorWidthInt", device.preferredVectorWidthInt());
-
-        DUMP_MESSAGE_STDOUT("    Preferred vector width long = "<< device.preferredVectorWidthLong());
-        DUMP_PROPERTY_XML("cv_ocl_current_preferredVectorWidthLong", device.preferredVectorWidthLong());
-
-        DUMP_MESSAGE_STDOUT("    Preferred vector width float = "<< device.preferredVectorWidthFloat());
-        DUMP_PROPERTY_XML("cv_ocl_current_preferredVectorWidthFloat", device.preferredVectorWidthFloat());
-
-        DUMP_MESSAGE_STDOUT("    Preferred vector width double = "<< device.preferredVectorWidthDouble());
-        DUMP_PROPERTY_XML("cv_ocl_current_preferredVectorWidthDouble", device.preferredVectorWidthDouble());
-    }
-    catch (...)
-    {
-        DUMP_MESSAGE_STDOUT("Exception. Can't dump OpenCL info");
-        DUMP_MESSAGE_STDOUT("OpenCL device not available");
-        DUMP_PROPERTY_XML("cv_ocl", "not available");
-    }
+    cv::dumpOpenCLInformation();
 }
-#undef DUMP_MESSAGE_STDOUT
-#undef DUMP_PROPERTY_XML
-
-#endif
+#endif // HAVE_OPENCL
 
 Mat TestUtils::readImage(const String &fileName, int flags)
 {
